+++ src/routes/lifi/utils.ts
@@ -10,19 +10,13 @@ import type {
   TokenId,
   TransactionId,
   Network,
-  Signer,
 } from '@wormhole-foundation/sdk-connect';
 import {
   Wormhole,
   chainToPlatform,
-  isSignAndSendSigner,
-  isSignOnlySigner,
   isNative,
   nativeChainIds,
 } from '@wormhole-foundation/sdk-connect';
-import type { EvmPlatform } from '@wormhole-foundation/sdk-evm';
-import type { SolanaPlatform } from '@wormhole-foundation/sdk-solana';
-import type { SuiPlatform } from '@wormhole-foundation/sdk-sui';
 import {
   LIFI_NATIVE_ADDRESS_EVM,
   LIFI_NATIVE_ADDRESS_SVM,
@@ -152,22 +146,4 @@ export async function getTransactionStatus(
     return null;
   }
   return response;
-}
-
-export async function executeTransaction<N extends Network>(
-  txReq: any,
-  signer: Signer<N>,
-  rpc: any,
-  chain: Chain,
-  platform: typeof SolanaPlatform | typeof SuiPlatform | typeof EvmPlatform,
-  txs: TransactionId[],
-): Promise<void> {
-  if (isSignAndSendSigner(signer)) {
-    const txids = await signer.signAndSend([txReq]);
-    txs.push(...txids.map((txid) => ({ chain, txid })));
-  } else if (isSignOnlySigner(signer)) {
-    const signed = await signer.sign([txReq]);
-    const txids = await platform.sendWait(chain, rpc, signed);
-    txs.push(...txids.map((txid) => ({ chain, txid })));
-  }
 }
